tsm1: move code around into smaller files and add tests
