+++ tsdb/tsm1/reader_test.go
@@ -1,18 +1,24 @@
 package tsm1
 
 import (
-	"fmt"
 	"io/ioutil"
 	"math"
 	"os"
 	"path/filepath"
 	"sort"
-	"sync/atomic"
 	"testing"
 )
 
-func fatal(t *testing.T, msg string, err error) {
+func fatal(t testing.TB, msg string, err error) {
+	t.Helper()
 	t.Fatalf("unexpected error %v: %v", msg, err)
+}
+
+func fatalIfErr(t testing.TB, msg string, err error) {
+	t.Helper()
+	if err != nil {
+		fatal(t, msg, err)
+	}
 }
 
 func TestTSMReader_Type(t *testing.T) {
@@ -54,6 +60,17 @@ func TestTSMReader_Type(t *testing.T) {
 
 	if got, exp := typ, BlockInteger; got != exp {
 		t.Fatalf("type mismatch: got %v, exp %v", got, exp)
+	}
+}
+
+func TestIndexWriter_MaxBlocks(t *testing.T) {
+	index := NewIndexWriter()
+	for i := 0; i < 1<<16; i++ {
+		index.Add([]byte("cpu"), BlockFloat64, 0, 1, 10, 20)
+	}
+
+	if _, err := index.MarshalBinary(); err == nil {
+		t.Fatalf("expected max block count error. got nil")
 	}
 }
 
@@ -1118,71 +1135,6 @@ func TestIndirectIndex_Entries(t *testing.T) {
 	}
 }
 
-func TestIndirectIndex_Entries_NonExistent(t *testing.T) {
-	index := NewIndexWriter()
-	index.Add([]byte("cpu"), BlockFloat64, 0, 1, 10, 100)
-	index.Add([]byte("cpu"), BlockFloat64, 2, 3, 20, 200)
-
-	b, err := index.MarshalBinary()
-	if err != nil {
-		t.Fatalf("unexpected error marshaling index: %v", err)
-	}
-
-	indirect := NewIndirectIndex()
-	if err := indirect.UnmarshalBinary(b); err != nil {
-		t.Fatalf("unexpected error unmarshaling index: %v", err)
-	}
-
-	// mem has not been added to the index so we should get no entries back
-	// for both
-	exp := index.Entries([]byte("mem"))
-	entries, err := indirect.ReadEntries([]byte("mem"), nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if got, exp := len(entries), len(exp); got != exp && exp != 0 {
-		t.Fatalf("entries length mismatch: got %v, exp %v", got, exp)
-	}
-}
-
-func TestIndirectIndex_MaxBlocks(t *testing.T) {
-	index := NewIndexWriter()
-	for i := 0; i < 1<<16; i++ {
-		index.Add([]byte("cpu"), BlockFloat64, 0, 1, 10, 20)
-	}
-
-	if _, err := index.MarshalBinary(); err == nil {
-		t.Fatalf("expected max block count error. got nil")
-	} else {
-		println(err.Error())
-	}
-}
-
-func TestIndirectIndex_Type(t *testing.T) {
-	index := NewIndexWriter()
-	index.Add([]byte("cpu"), BlockInteger, 0, 1, 10, 20)
-
-	b, err := index.MarshalBinary()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ind := NewIndirectIndex()
-	if err := ind.UnmarshalBinary(b); err != nil {
-		fatal(t, "unmarshal binary", err)
-	}
-
-	typ, err := ind.Type([]byte("cpu"))
-	if err != nil {
-		fatal(t, "reading type", err)
-	}
-
-	if got, exp := typ, BlockInteger; got != exp {
-		t.Fatalf("type mismatch: got %v, exp %v", got, exp)
-	}
-}
-
 func TestDirectIndex_KeyCount(t *testing.T) {
 	index := NewIndexWriter()
 	index.Add([]byte("cpu"), BlockFloat64, 0, 1, 10, 20)
@@ -1195,280 +1147,7 @@ func TestDirectIndex_KeyCount(t *testing.T) {
 	}
 }
 
-func TestBlockIterator_Single(t *testing.T) {
-	dir := mustTempDir()
-	defer os.RemoveAll(dir)
-	f := mustTempFile(dir)
-
-	w, err := NewTSMWriter(f)
-	if err != nil {
-		t.Fatalf("unexpected error creating writer: %v", err)
-	}
-
-	values := []Value{NewValue(0, int64(1))}
-	if err := w.Write([]byte("cpu"), values); err != nil {
-		t.Fatalf("unexpected error writing: %v", err)
-
-	}
-	if err := w.WriteIndex(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	if err := w.Close(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	fd, err := os.Open(f.Name())
-	if err != nil {
-		t.Fatalf("unexpected error opening: %v", err)
-	}
-
-	r, err := NewTSMReader(fd)
-	if err != nil {
-		t.Fatalf("unexpected error created reader: %v", err)
-	}
-
-	var count int
-	iter := r.BlockIterator()
-	for iter.Next() {
-		key, minTime, maxTime, typ, _, buf, err := iter.Read()
-		if err != nil {
-			t.Fatalf("unexpected error creating iterator: %v", err)
-		}
-
-		if got, exp := string(key), "cpu"; got != exp {
-			t.Fatalf("key mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := minTime, int64(0); got != exp {
-			t.Fatalf("min time mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := maxTime, int64(0); got != exp {
-			t.Fatalf("max time mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := typ, BlockInteger; got != exp {
-			t.Fatalf("block type mismatch: got %v, exp %v", got, exp)
-		}
-
-		if len(buf) == 0 {
-			t.Fatalf("buf length = 0")
-		}
-
-		count++
-	}
-
-	if got, exp := count, len(values); got != exp {
-		t.Fatalf("value count mismatch: got %v, exp %v", got, exp)
-	}
-}
-
-func TestBlockIterator_Tombstone(t *testing.T) {
-	dir := mustTempDir()
-	defer os.RemoveAll(dir)
-	f := mustTempFile(dir)
-
-	w, err := NewTSMWriter(f)
-	if err != nil {
-		t.Fatalf("unexpected error creating writer: %v", err)
-	}
-
-	values := []Value{NewValue(0, int64(1))}
-	if err := w.Write([]byte("cpu"), values); err != nil {
-		t.Fatalf("unexpected error writing: %v", err)
-	}
-
-	if err := w.Write([]byte("mem"), values); err != nil {
-		t.Fatalf("unexpected error writing: %v", err)
-	}
-
-	if err := w.WriteIndex(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	if err := w.Close(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	fd, err := os.Open(f.Name())
-	if err != nil {
-		t.Fatalf("unexpected error opening: %v", err)
-	}
-
-	r, err := NewTSMReader(fd)
-	if err != nil {
-		t.Fatalf("unexpected error created reader: %v", err)
-	}
-
-	iter := r.BlockIterator()
-	for iter.Next() {
-		// Trigger a delete during iteration.  This should cause an error condition for
-		// the BlockIterator
-		r.Delete([][]byte{[]byte("cpu")})
-	}
-
-	if iter.Err() == nil {
-		t.Fatalf("expected error: got nil")
-	}
-}
-
-func TestBlockIterator_MultipleBlocks(t *testing.T) {
-	dir := mustTempDir()
-	defer os.RemoveAll(dir)
-	f := mustTempFile(dir)
-
-	w, err := NewTSMWriter(f)
-	if err != nil {
-		t.Fatalf("unexpected error creating writer: %v", err)
-	}
-
-	values1 := []Value{NewValue(0, int64(1))}
-	if err := w.Write([]byte("cpu"), values1); err != nil {
-		t.Fatalf("unexpected error writing: %v", err)
-	}
-
-	values2 := []Value{NewValue(1, int64(2))}
-	if err := w.Write([]byte("cpu"), values2); err != nil {
-		t.Fatalf("unexpected error writing: %v", err)
-	}
-
-	if err := w.WriteIndex(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	if err := w.Close(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	fd, err := os.Open(f.Name())
-	if err != nil {
-		t.Fatalf("unexpected error opening: %v", err)
-	}
-
-	r, err := NewTSMReader(fd)
-	if err != nil {
-		t.Fatalf("unexpected error created reader: %v", err)
-	}
-
-	var count int
-	expData := []Values{values1, values2}
-	iter := r.BlockIterator()
-	var i int
-	for iter.Next() {
-		key, minTime, maxTime, typ, _, buf, err := iter.Read()
-
-		if err != nil {
-			t.Fatalf("unexpected error creating iterator: %v", err)
-		}
-
-		if got, exp := string(key), "cpu"; got != exp {
-			t.Fatalf("key mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := minTime, expData[i][0].UnixNano(); got != exp {
-			t.Fatalf("min time mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := maxTime, expData[i][0].UnixNano(); got != exp {
-			t.Fatalf("max time mismatch: got %v, exp %v", got, exp)
-		}
-
-		if got, exp := typ, BlockInteger; got != exp {
-			t.Fatalf("block type mismatch: got %v, exp %v", got, exp)
-		}
-
-		if len(buf) == 0 {
-			t.Fatalf("buf length = 0")
-		}
-
-		count++
-		i++
-	}
-
-	if got, exp := count, 2; got != exp {
-		t.Fatalf("value count mismatch: got %v, exp %v", got, exp)
-	}
-}
-
-func TestBlockIterator_Sorted(t *testing.T) {
-	dir := mustTempDir()
-	defer os.RemoveAll(dir)
-	f := mustTempFile(dir)
-
-	w, err := NewTSMWriter(f)
-	if err != nil {
-		t.Fatalf("unexpected error creating writer: %v", err)
-	}
-
-	values := map[string][]Value{
-		"mem":    []Value{NewValue(0, int64(1))},
-		"cycles": []Value{NewValue(0, ^uint64(0))},
-		"cpu":    []Value{NewValue(1, float64(2))},
-		"disk":   []Value{NewValue(1, true)},
-		"load":   []Value{NewValue(1, "string")},
-	}
-
-	keys := make([]string, 0, len(values))
-	for k := range values {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	for _, k := range keys {
-		if err := w.Write([]byte(k), values[k]); err != nil {
-			t.Fatalf("unexpected error writing: %v", err)
-
-		}
-	}
-
-	if err := w.WriteIndex(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	if err := w.Close(); err != nil {
-		t.Fatalf("unexpected error closing: %v", err)
-	}
-
-	fd, err := os.Open(f.Name())
-	if err != nil {
-		t.Fatalf("unexpected error opening: %v", err)
-	}
-
-	r, err := NewTSMReader(fd)
-	if err != nil {
-		t.Fatalf("unexpected error created reader: %v", err)
-	}
-
-	var count int
-	iter := r.BlockIterator()
-	var lastKey string
-	for iter.Next() {
-		key, _, _, _, _, buf, err := iter.Read()
-
-		if string(key) < lastKey {
-			t.Fatalf("keys not sorted: got %v, last %v", key, lastKey)
-		}
-
-		lastKey = string(key)
-
-		if err != nil {
-			t.Fatalf("unexpected error creating iterator: %v", err)
-		}
-
-		if len(buf) == 0 {
-			t.Fatalf("buf length = 0")
-		}
-
-		count++
-	}
-
-	if got, exp := count, len(values); got != exp {
-		t.Fatalf("value count mismatch: got %v, exp %v", got, exp)
-	}
-}
-
-func TestIndirectIndex_UnmarshalBinary_BlockCountOverflow(t *testing.T) {
+func TestTSMReader_UnmarshalBinary_BlockCountOverflow(t *testing.T) {
 	dir := mustTempDir()
 	defer os.RemoveAll(dir)
 	f := mustTempFile(dir)
@@ -1502,6 +1181,7 @@ func TestIndirectIndex_UnmarshalBinary_BlockCountOverflow(t *testing.T) {
 	}
 	defer r.Close()
 }
+
 
 func TestCompacted_NotFull(t *testing.T) {
 	dir := mustTempDir()
@@ -1870,317 +1550,4 @@ func TestTSMReader_References(t *testing.T) {
 	if err := r.Remove(); err != nil {
 		t.Fatalf("unexpected error removing reader: %v", err)
 	}
-}
-
-//
-// indirectIndex benchmarks
-//
-
-const (
-	indexKeyCount   = 500000
-	indexBlockCount = 100
-)
-
-type indexCacheInfo struct {
-	index    *indirectIndex
-	offsets  []uint32
-	prefixes []prefixEntry
-	allKeys  [][]byte
-	bytes    []byte
-}
-
-func (i *indexCacheInfo) reset() {
-	i.index.ro.offsets = append([]uint32(nil), i.offsets...)
-	i.index.ro.prefixes = append([]prefixEntry(nil), i.prefixes...)
-	i.index.tombstones = make(map[uint32][]TimeRange)
-	i.index.prefixTombstones = newPrefixTree()
-	resetFaults(i.index)
-}
-
-var (
-	indexCache = map[string]*indexCacheInfo{}
-	indexSizes = map[string][2]int{
-		"large": {500000, 100},
-		"med":   {1000, 1000},
-		"small": {5000, 2},
-	}
-)
-
-func getFaults(indirect *indirectIndex) int64 {
-	return int64(atomic.LoadUint64(&indirect.b.faults))
-}
-
-func resetFaults(indirect *indirectIndex) {
-	if indirect != nil {
-		indirect.b = faultBuffer{b: indirect.b.b}
-	}
-}
-
-func getIndex(tb testing.TB, name string) (*indirectIndex, *indexCacheInfo) {
-	info, ok := indexCache[name]
-	if ok {
-		info.reset()
-		return info.index, info
-	}
-	info = new(indexCacheInfo)
-
-	sizes, ok := indexSizes[name]
-	if !ok {
-		sizes = [2]int{indexKeyCount, indexBlockCount}
-	}
-	keys, blocks := sizes[0], sizes[1]
-
-	writer := NewIndexWriter()
-
-	// add a ballast key that starts at -1 so that we don't trigger optimizations
-	// when deleting [0, MaxInt]
-	writer.Add([]byte("ballast"), BlockFloat64, -1, 1, 0, 100)
-
-	for i := 0; i < keys; i++ {
-		key := []byte(fmt.Sprintf("cpu-%08d", i))
-		info.allKeys = append(info.allKeys, key)
-		for j := 0; j < blocks; j++ {
-			writer.Add(key, BlockFloat64, 0, 100, 10, 100)
-		}
-	}
-
-	var err error
-	info.bytes, err = writer.MarshalBinary()
-	if err != nil {
-		tb.Fatalf("unexpected error marshaling index: %v", err)
-	}
-
-	info.index = NewIndirectIndex()
-	if err = info.index.UnmarshalBinary(info.bytes); err != nil {
-		tb.Fatalf("unexpected error unmarshaling index: %v", err)
-	}
-	info.offsets = append([]uint32(nil), info.index.ro.offsets...)
-	info.prefixes = append([]prefixEntry(nil), info.index.ro.prefixes...)
-
-	indexCache[name] = info
-	return info.index, info
-}
-
-func BenchmarkIndirectIndex_UnmarshalBinary(b *testing.B) {
-	indirect, info := getIndex(b, "large")
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		if err := indirect.UnmarshalBinary(info.bytes); err != nil {
-			b.Fatalf("unexpected error unmarshaling index: %v", err)
-		}
-	}
-}
-
-func BenchmarkIndirectIndex_Entries(b *testing.B) {
-	indirect, _ := getIndex(b, "med")
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		resetFaults(indirect)
-		indirect.ReadEntries([]byte("cpu-00000001"), nil)
-	}
-
-	if faultBufferEnabled {
-		b.SetBytes(getFaults(indirect) * 4096)
-		b.Log("recorded faults:", getFaults(indirect))
-	}
-}
-
-func BenchmarkIndirectIndex_ReadEntries(b *testing.B) {
-	var entries []IndexEntry
-	indirect, _ := getIndex(b, "med")
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		resetFaults(indirect)
-		entries, _ = indirect.ReadEntries([]byte("cpu-00000001"), entries)
-	}
-
-	if faultBufferEnabled {
-		b.SetBytes(getFaults(indirect) * 4096)
-		b.Log("recorded faults:", getFaults(indirect))
-	}
-}
-
-func BenchmarkBlockIterator_Next(b *testing.B) {
-	indirect, _ := getIndex(b, "med")
-	r := TSMReader{index: indirect}
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		resetFaults(indirect)
-		bi := r.BlockIterator()
-		for bi.Next() {
-		}
-	}
-
-	if faultBufferEnabled {
-		b.SetBytes(getFaults(indirect) * 4096)
-		b.Log("recorded faults:", getFaults(indirect))
-	}
-}
-
-func BenchmarkIndirectIndex_DeleteRangeLast(b *testing.B) {
-	indirect, _ := getIndex(b, "large")
-	keys := [][]byte{[]byte("cpu-00999999")}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		resetFaults(indirect)
-		indirect.DeleteRange(keys, 10, 50)
-	}
-
-	if faultBufferEnabled {
-		b.SetBytes(getFaults(indirect) * 4096)
-		b.Log("recorded faults:", getFaults(indirect))
-	}
-}
-
-func BenchmarkIndirectIndex_DeleteRangeFull(b *testing.B) {
-	run := func(b *testing.B, name string) {
-		indirect, _ := getIndex(b, name)
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for i := 0; i < b.N; i++ {
-			b.StopTimer()
-			var info *indexCacheInfo
-			indirect, info = getIndex(b, name)
-			b.StartTimer()
-
-			for i := 0; i < len(info.allKeys); i += 4096 {
-				n := i + 4096
-				if n > len(info.allKeys) {
-					n = len(info.allKeys)
-				}
-				indirect.DeleteRange(info.allKeys[i:n], 10, 50)
-			}
-		}
-
-		if faultBufferEnabled {
-			b.SetBytes(getFaults(indirect) * 4096)
-			b.Log("recorded faults:", getFaults(indirect))
-		}
-	}
-
-	b.Run("Large", func(b *testing.B) { run(b, "large") })
-	b.Run("Small", func(b *testing.B) { run(b, "small") })
-}
-
-func BenchmarkIndirectIndex_DeleteRangeFull_Covered(b *testing.B) {
-	run := func(b *testing.B, name string) {
-		indirect, _ := getIndex(b, name)
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for i := 0; i < b.N; i++ {
-			b.StopTimer()
-			var info *indexCacheInfo
-			indirect, info = getIndex(b, name)
-			b.StartTimer()
-
-			for i := 0; i < len(info.allKeys); i += 4096 {
-				n := i + 4096
-				if n > len(info.allKeys) {
-					n = len(info.allKeys)
-				}
-				indirect.DeleteRange(info.allKeys[i:n], 0, math.MaxInt64)
-			}
-		}
-
-		if faultBufferEnabled {
-			b.SetBytes(getFaults(indirect) * 4096)
-			b.Log("recorded faults:", getFaults(indirect))
-		}
-	}
-
-	b.Run("Large", func(b *testing.B) { run(b, "large") })
-	b.Run("Small", func(b *testing.B) { run(b, "small") })
-}
-
-func BenchmarkIndirectIndex_Delete(b *testing.B) {
-	run := func(b *testing.B, name string) {
-		indirect, _ := getIndex(b, name)
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for i := 0; i < b.N; i++ {
-			b.StopTimer()
-			var info *indexCacheInfo
-			indirect, info = getIndex(b, name)
-			b.StartTimer()
-
-			for i := 0; i < len(info.allKeys); i += 4096 {
-				n := i + 4096
-				if n > len(info.allKeys) {
-					n = len(info.allKeys)
-				}
-				indirect.Delete(info.allKeys[i:n])
-			}
-		}
-
-		if faultBufferEnabled {
-			b.SetBytes(getFaults(indirect) * 4096)
-			b.Log("recorded faults:", getFaults(indirect))
-		}
-	}
-
-	b.Run("Large", func(b *testing.B) { run(b, "large") })
-	b.Run("Small", func(b *testing.B) { run(b, "small") })
-}
-
-func BenchmarkIndirectIndex_DeletePrefixFull(b *testing.B) {
-	prefix := []byte("cpu-")
-	run := func(b *testing.B, name string) {
-		indirect, _ := getIndex(b, name)
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for i := 0; i < b.N; i++ {
-			b.StopTimer()
-			indirect, _ = getIndex(b, name)
-			b.StartTimer()
-
-			indirect.DeletePrefix(prefix, 10, 50)
-		}
-
-		if faultBufferEnabled {
-			b.SetBytes(getFaults(indirect) * 4096)
-			b.Log("recorded faults:", getFaults(indirect))
-		}
-	}
-
-	b.Run("Large", func(b *testing.B) { run(b, "large") })
-	b.Run("Small", func(b *testing.B) { run(b, "small") })
-}
-
-func BenchmarkIndirectIndex_DeletePrefixFull_Covered(b *testing.B) {
-	prefix := []byte("cpu-")
-	run := func(b *testing.B, name string) {
-		indirect, _ := getIndex(b, name)
-		b.ReportAllocs()
-		b.ResetTimer()
-
-		for i := 0; i < b.N; i++ {
-			b.StopTimer()
-			indirect, _ = getIndex(b, name)
-			b.StartTimer()
-
-			indirect.DeletePrefix(prefix, 0, math.MaxInt64)
-		}
-
-		if faultBufferEnabled {
-			b.SetBytes(getFaults(indirect) * 4096)
-			b.Log("recorded faults:", getFaults(indirect))
-		}
-	}
-
-	b.Run("Large", func(b *testing.B) { run(b, "large") })
-	b.Run("Small", func(b *testing.B) { run(b, "small") })
 }
